requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cssselect==1.1.0
fake-useragent==0.1.11
h11==0.12.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.2
lxml==4.6.3
outcome==1.1.0
parse==1.19.0
psycopg2-binary==2.9.2
pycparser==2.21
pyee==8.2.2
pymongo==3.12.1
pyppeteer==0.2.6
pyquery==1.4.3
requests==2.26.0
requests-html==0.10.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
w3lib==1.22.0
webencodings==0.5.1
websockets==9.1
wsproto==1.0.0
zipp==3.6.0
